refactor: move plot de confusion matrix plotting para um novo modulo
